+++ programs/dynamic-bonding-curve/src/instructions/migration/dynamic_amm_v2/migrate_damm_v2_initialize_pool.rs
@@ -4,31 +4,31 @@ use anchor_spl::{
     token_2022::{set_authority, spl_token_2022::instruction::AuthorityType, SetAuthority},
     token_interface::{TokenAccount, TokenInterface},
 };
-use damm_v2::types::{
-    AddLiquidityParameters, BaseFeeParameters, InitializeCustomizablePoolParameters,
-    InitializePoolParameters, PoolFeeParameters,
+use damm_v2::{
+    accounts::PodAlignedFeeTimeScheduler,
+    types::{
+        AddLiquidityParameters, InitializeCustomizablePoolParameters, InitializePoolParameters,
+    },
 };
 use ruint::aliases::U512;
 
 use crate::{
     activation_handler::ActivationType,
-    calculate_dynamic_fee_params,
     const_pda::{self, pool_authority::BUMP},
-    constants::{
-        fee::FEE_DENOMINATOR, seeds::POSITION_VESTING_PREFIX, MAX_SQRT_PRICE, MIN_SQRT_PRICE,
-    },
+    constants::{seeds::POSITION_VESTING_PREFIX, MAX_SQRT_PRICE, MIN_SQRT_PRICE},
     convert_collect_fee_mode_to_dammv2,
     cpi_checker::cpi_with_account_lamport_and_owner_checking,
     curve::{get_initial_liquidity_from_delta_base, get_initial_liquidity_from_delta_quote},
-    flash_rent,
-    params::fee_parameters::{to_bps, to_numerator},
+    damm_v2_utils, flash_rent,
+    params::{fee_parameters::to_bps, liquidity_distribution::get_protocol_migration_fee},
     safe_math::SafeMath,
     state::{
         LiquidityDistribution, LiquidityDistributionItem, MigrationAmount, MigrationFeeOption,
         MigrationOption, MigrationProgress, PoolConfig, VirtualPool,
     },
-    DammV2DynamicFee, PoolError,
+    PoolError,
 };
+use damm_v2_utils::BaseFeeMode as DammV2BaseFeeMode;
 
 #[derive(Accounts)]
 pub struct MigrateDammV2Ctx<'info> {
@@ -149,34 +149,11 @@ impl<'info> MigrateDammV2Ctx<'info> {
                 self.system_program.to_account_info(),
                 || {
                     if migration_fee_option == MigrationFeeOption::Customizable {
-                        let base_fee_numerator = to_numerator(
-                            config.migrated_pool_fee_bps.into(),
-                            FEE_DENOMINATOR.into(),
-                        )?;
-                        let base_fee = BaseFeeParameters {
-                            cliff_fee_numerator: base_fee_numerator,
-                            ..Default::default()
-                        };
-
-                        let migrated_dynamic_fee =
-                            DammV2DynamicFee::try_from(config.migrated_dynamic_fee)
-                                .map_err(|_| PoolError::InvalidCollectFeeMode)?;
-
-                        let dynamic_fee_params = if migrated_dynamic_fee == DammV2DynamicFee::Enable
-                        {
-                            Some(calculate_dynamic_fee_params(base_fee_numerator)?)
-                        } else {
-                            None
-                        };
-
-                        let pool_fees = PoolFeeParameters {
-                            base_fee,
-                            padding: [0; 3],
-                            dynamic_fee: dynamic_fee_params,
-                        };
+                        let pool_fees = config.build_damm_v2_pool_fee_params()?;
 
                         let collect_fee_mode =
                             convert_collect_fee_mode_to_dammv2(config.migrated_collect_fee_mode)?;
+
                         let initialize_pool_params = InitializeCustomizablePoolParameters {
                             pool_fees,
                             sqrt_min_price: MIN_SQRT_PRICE,
@@ -481,21 +458,36 @@ fn validate_config_key(
         | MigrationFeeOption::FixedBps200
         | MigrationFeeOption::FixedBps400
         | MigrationFeeOption::FixedBps600 => {
-            let base_fee_bps = to_bps(
-                damm_config.pool_fees.base_fee.cliff_fee_numerator.into(),
-                1_000_000_000, // damm v2 using the same fee denominator with virtual curve
-            )?;
+            let fee_scheduler = bytemuck::try_from_bytes::<PodAlignedFeeTimeScheduler>(
+                &damm_config.pool_fees.base_fee.data,
+            )
+            .map_err(|_| PoolError::UndeterminedError)?;
+
+            let base_fee_mode: DammV2BaseFeeMode = fee_scheduler
+                .base_fee_mode
+                .try_into()
+                .map_err(|_| PoolError::TypeCastFailed)?;
 
             // Validate it's fee scheduler linear | exponential
             require!(
-                damm_config.pool_fees.base_fee.base_fee_mode < 2,
+                base_fee_mode == DammV2BaseFeeMode::FeeTimeSchedulerLinear
+                    || base_fee_mode == DammV2BaseFeeMode::FeeTimeSchedulerExponential,
                 PoolError::InvalidConfigAccount
             );
 
-            let period_frequency = u64::from_le_bytes(damm_config.pool_fees.base_fee.second_factor);
+            let base_fee_bps = to_bps(
+                fee_scheduler.cliff_fee_numerator.into(),
+                damm_v2::constants::FEE_DENOMINATOR.into(),
+            )?;
 
             // Validate no schedule
-            require!(period_frequency == 0, PoolError::InvalidConfigAccount);
+            require!(
+                fee_scheduler.period_frequency == 0
+                    && fee_scheduler.reduction_factor == 0
+                    && fee_scheduler.number_of_period == 0,
+                PoolError::InvalidConfigAccount
+            );
+
             migration_fee_option.validate_base_fee(base_fee_bps)?;
 
             require!(
@@ -582,13 +574,30 @@ pub fn handle_migrate_damm_v2<'c: 'info, 'info>(
     let migration_sqrt_price = config.migration_sqrt_price;
 
     let MigrationAmount { quote_amount, .. } = config.get_migration_quote_amount_for_config()?;
+    // Migration base threshold + buffer
     let excluded_fee_base_reserve =
         initial_base_vault_amount.safe_sub(protocol_and_partner_base_fee)?;
 
-    // calculate initial liquidity
-    let initial_liquidity = get_liquidity_for_adding_liquidity(
+    let (protocol_migration_base_fee, protocol_migration_quote_fee) = get_protocol_migration_fee(
         excluded_fee_base_reserve,
         quote_amount,
+        migration_sqrt_price,
+        virtual_pool.protocol_liquidity_migration_fee_bps,
+        MigrationOption::DammV2,
+    )?;
+
+    virtual_pool.save_protocol_liquidity_migration_fee(
+        protocol_migration_base_fee,
+        protocol_migration_quote_fee,
+    );
+
+    let migration_base_amount = excluded_fee_base_reserve.safe_sub(protocol_migration_base_fee)?;
+    let migration_quote_amount = quote_amount.safe_sub(protocol_migration_quote_fee)?;
+
+    // calculate initial liquidity
+    let initial_liquidity = get_liquidity_for_adding_liquidity(
+        migration_base_amount,
+        migration_quote_amount,
         migration_sqrt_price,
     )?;
 
@@ -659,13 +668,14 @@ pub fn handle_migrate_damm_v2<'c: 'info, 'info>(
     let deposited_quote_amount =
         initial_quote_vault_amount.safe_sub(ctx.accounts.quote_vault.amount)?;
 
-    let updated_excluded_fee_base_reserve =
-        excluded_fee_base_reserve.safe_sub(deposited_base_amount)?;
-    let updated_quote_threshold = quote_amount.safe_sub(deposited_quote_amount)?;
+    let leftover_migration_base_amount = migration_base_amount.safe_sub(deposited_base_amount)?;
+
+    let leftover_migration_quote_amount =
+        migration_quote_amount.safe_sub(deposited_quote_amount)?;
 
     let liquidity_for_second_position = get_liquidity_for_adding_liquidity(
-        updated_excluded_fee_base_reserve,
-        updated_quote_threshold,
+        leftover_migration_base_amount,
+        leftover_migration_quote_amount,
         migration_sqrt_price,
     )?;
 
@@ -719,12 +729,14 @@ pub fn handle_migrate_damm_v2<'c: 'info, 'info>(
     ctx.accounts.base_vault.reload()?;
 
     // check whether we should burn token
+    let non_burnable_amount =
+        protocol_and_partner_base_fee.safe_add(protocol_migration_base_fee)?;
 
     let left_base_token = ctx
         .accounts
         .base_vault
         .amount
-        .safe_sub(protocol_and_partner_base_fee)?;
+        .safe_sub(non_burnable_amount)?;
 
     let burnable_amount = config.get_burnable_amount_post_migration(left_base_token)?;
 
@@ -751,7 +763,7 @@ pub fn handle_migrate_damm_v2<'c: 'info, 'info>(
     Ok(())
 }
 
-fn get_liquidity_for_adding_liquidity(
+pub(crate) fn get_liquidity_for_adding_liquidity(
     base_amount: u64,
     quote_amount: u64,
     sqrt_price: u128,