+++ programs/dynamic-bonding-curve/src/instructions/migration/withdraw_leftover.rs
@@ -76,13 +76,14 @@ pub fn handle_withdraw_leftover(ctx: Context<WithdrawLeftoverCtx>) -> Result<()>
         .accounts
         .base_vault
         .amount
-        .safe_sub(virtual_pool.get_protocol_and_trading_base_fee()?)?;
+        .safe_sub(virtual_pool.get_protocol_and_trading_base_fee()?)?
+        .safe_sub(virtual_pool.protocol_migration_base_fee_amount)?;
 
     transfer_token_from_pool_authority(
         ctx.accounts.pool_authority.to_account_info(),
         &ctx.accounts.base_mint,
         &ctx.accounts.base_vault,
-        &ctx.accounts.token_base_account,
+        ctx.accounts.token_base_account.to_account_info(),
         &ctx.accounts.token_base_program,
         leftover_amount,
     )?;