+++ programs/dynamic-bonding-curve/src/state/virtual_pool.rs
@@ -132,7 +132,9 @@ pub struct VirtualPool {
     pub is_withdraw_leftover: u8,
     /// is creator withdraw surplus
     pub is_creator_withdraw_surplus: u8,
-    /// migration fee withdraw status, first bit is for partner, second bit is for creator
+    /// migration fee withdraw status
+    /// bit 1 (0b010) creator
+    /// bit 2 (0b100) partner
     pub migration_fee_withdraw_status: u8,
     /// pool metrics
     pub metrics: PoolMetrics,
@@ -146,10 +148,16 @@ pub struct VirtualPool {
     pub legacy_creation_fee_bits: u8,
     /// pool creation fee claim status
     pub creation_fee_bits: u8,
+    /// Cached flag
+    pub has_swap: u8,
     /// Padding for further use
-    pub _padding_0: [u8; 6],
+    pub _padding_0: [u8; 5],
+    pub protocol_liquidity_migration_fee_bps: u16,
+    pub _padding_1: [u8; 6],
+    pub protocol_migration_base_fee_amount: u64,
+    pub protocol_migration_quote_fee_amount: u64,
     /// Padding for further use
-    pub _padding_1: [u64; 6],
+    pub _padding_2: [u64; 3],
 }
 
 const_assert_eq!(VirtualPool::INIT_SPACE, 416);
@@ -206,6 +214,7 @@ impl VirtualPool {
         pool_type: u8,
         activation_point: u64,
         base_reserve: u64,
+        protocol_liquidity_migration_fee_bps: u16,
     ) {
         self.volatility_tracker = volatility_tracker;
         self.config = config;
@@ -217,6 +226,7 @@ impl VirtualPool {
         self.pool_type = pool_type;
         self.activation_point = activation_point;
         self.base_reserve = base_reserve;
+        self.protocol_liquidity_migration_fee_bps = protocol_liquidity_migration_fee_bps;
     }
 
     pub fn get_swap_result_from_exact_output(
@@ -226,6 +236,7 @@ impl VirtualPool {
         fee_mode: &FeeMode,
         trade_direction: TradeDirection,
         current_point: u64,
+        eligible_for_first_swap_with_min_fee: bool,
     ) -> Result<SwapResult2> {
         let mut actual_protocol_fee = 0;
         let mut actual_trading_fee = 0;
@@ -234,15 +245,20 @@ impl VirtualPool {
         let included_fee_out_amount = if fee_mode.fees_on_input {
             amount_out
         } else {
-            let trade_fee_numerator = config
-                .pool_fees
-                .get_total_fee_numerator_from_excluded_fee_amount(
-                    &self.volatility_tracker,
-                    current_point,
-                    self.activation_point,
-                    amount_out,
-                    trade_direction,
-                )?;
+            let trade_fee_numerator = if eligible_for_first_swap_with_min_fee {
+                config.pool_fees.get_min_base_fee_numerator()?
+            } else {
+                config
+                    .pool_fees
+                    .get_total_fee_numerator_from_excluded_fee_amount(
+                        &self.volatility_tracker,
+                        current_point,
+                        self.activation_point,
+                        amount_out,
+                        trade_direction,
+                    )?
+            };
+
             let (included_fee_out_amount, fee_amount) =
                 PoolFeesConfig::get_included_fee_amount(trade_fee_numerator, amount_out)?;
 
@@ -271,19 +287,23 @@ impl VirtualPool {
 
         require!(
             next_sqrt_price <= config.migration_sqrt_price,
-            PoolError::SwapAmountIsOverAThreshold
+            PoolError::InsufficientLiquidity
         );
 
         let (excluded_fee_input_amount, included_fee_input_amount) = if fee_mode.fees_on_input {
-            let trade_fee_numerator = config
-                .pool_fees
-                .get_total_fee_numerator_from_excluded_fee_amount(
-                    &self.volatility_tracker,
-                    current_point,
-                    self.activation_point,
-                    amount_in,
-                    trade_direction,
-                )?;
+            let trade_fee_numerator = if eligible_for_first_swap_with_min_fee {
+                config.pool_fees.get_min_base_fee_numerator()?
+            } else {
+                config
+                    .pool_fees
+                    .get_total_fee_numerator_from_excluded_fee_amount(
+                        &self.volatility_tracker,
+                        current_point,
+                        self.activation_point,
+                        amount_in,
+                        trade_direction,
+                    )?
+            };
 
             let (included_fee_in_amount, fee_amount) =
                 PoolFeesConfig::get_included_fee_amount(trade_fee_numerator, amount_in)?;
@@ -366,15 +386,26 @@ impl VirtualPool {
             }
         }
         if amount_left != 0 {
+            let max_amount_out = get_delta_amount_quote_unsigned_256(
+                config.sqrt_start_price,
+                current_sqrt_price,
+                config.curve[0].liquidity,
+                Rounding::Down,
+            )?;
+            require!(
+                U256::from(amount_left) <= max_amount_out,
+                PoolError::InsufficientLiquidity
+            );
             let next_sqrt_price = get_next_sqrt_price_from_output(
                 current_sqrt_price,
                 config.curve[0].liquidity,
                 amount_left,
                 true,
             )?;
+            // redundant check
             require!(
                 next_sqrt_price >= config.sqrt_start_price,
-                PoolError::NextSqrtPriceIsSmallerThanStartSqrtPrice
+                PoolError::InsufficientLiquidity
             );
             let in_amount = get_delta_amount_base_unsigned(
                 next_sqrt_price,
@@ -465,20 +496,25 @@ impl VirtualPool {
         fee_mode: &FeeMode,
         trade_direction: TradeDirection,
         current_point: u64,
+        eligible_for_first_swap_with_min_fee: bool,
     ) -> Result<SwapResult2> {
         let mut actual_protocol_fee = 0;
         let mut actual_trading_fee = 0;
         let mut actual_referral_fee = 0;
 
-        let trade_fee_numerator = config
-            .pool_fees
-            .get_total_fee_numerator_from_included_fee_amount(
-                &self.volatility_tracker,
-                current_point,
-                self.activation_point,
-                amount_in,
-                trade_direction,
-            )?;
+        let trade_fee_numerator = if eligible_for_first_swap_with_min_fee {
+            config.pool_fees.get_min_base_fee_numerator()?
+        } else {
+            config
+                .pool_fees
+                .get_total_fee_numerator_from_included_fee_amount(
+                    &self.volatility_tracker,
+                    current_point,
+                    self.activation_point,
+                    amount_in,
+                    trade_direction,
+                )?
+        };
 
         let actual_amount_in = if fee_mode.fees_on_input {
             let FeeOnAmountResult {
@@ -516,7 +552,7 @@ impl VirtualPool {
             )?,
         };
 
-        require!(amount_left == 0, PoolError::SwapAmountIsOverAThreshold);
+        require!(amount_left == 0, PoolError::InsufficientLiquidity);
 
         let actual_amount_out = if fee_mode.fees_on_input {
             output_amount
@@ -558,20 +594,25 @@ impl VirtualPool {
         fee_mode: &FeeMode,
         trade_direction: TradeDirection,
         current_point: u64,
+        eligible_for_first_swap_with_min_fee: bool,
     ) -> Result<SwapResult2> {
         let mut actual_protocol_fee = 0;
         let mut actual_trading_fee = 0;
         let mut actual_referral_fee = 0;
 
-        let trade_fee_numerator = config
-            .pool_fees
-            .get_total_fee_numerator_from_included_fee_amount(
-                &self.volatility_tracker,
-                current_point,
-                self.activation_point,
-                amount_in,
-                trade_direction,
-            )?;
+        let trade_fee_numerator = if eligible_for_first_swap_with_min_fee {
+            config.pool_fees.get_min_base_fee_numerator()?
+        } else {
+            config
+                .pool_fees
+                .get_total_fee_numerator_from_included_fee_amount(
+                    &self.volatility_tracker,
+                    current_point,
+                    self.activation_point,
+                    amount_in,
+                    trade_direction,
+                )?
+        };
 
         let mut actual_amount_in = if fee_mode.fees_on_input {
             let FeeOnAmountResult {
@@ -613,15 +654,20 @@ impl VirtualPool {
             actual_amount_in = actual_amount_in.safe_sub(amount_left)?;
             // recalculate included_fee_input_amount actual_trading_fee, actual_protocol_fee, actual_referral_fee
             if fee_mode.fees_on_input {
-                let trade_fee_numerator = config
-                    .pool_fees
-                    .get_total_fee_numerator_from_excluded_fee_amount(
-                        &self.volatility_tracker,
-                        current_point,
-                        self.activation_point,
-                        actual_amount_in,
-                        trade_direction,
-                    )?;
+                let trade_fee_numerator = if eligible_for_first_swap_with_min_fee {
+                    config.pool_fees.get_min_base_fee_numerator()?
+                } else {
+                    config
+                        .pool_fees
+                        .get_total_fee_numerator_from_excluded_fee_amount(
+                            &self.volatility_tracker,
+                            current_point,
+                            self.activation_point,
+                            actual_amount_in,
+                            trade_direction,
+                        )?
+                };
+
                 let (included_fee_input_amount, fee_amount) =
                     PoolFeesConfig::get_included_fee_amount(trade_fee_numerator, actual_amount_in)?;
 
@@ -694,7 +740,7 @@ impl VirtualPool {
                     config.curve[i].sqrt_price,
                     current_sqrt_price,
                     config.curve[i + 1].liquidity,
-                    Rounding::Up, // TODO check whether we should use round down or round up
+                    Rounding::Up,
                 )?;
                 if U256::from(amount_left) < max_amount_in {
                     let next_sqrt_price = get_next_sqrt_price_from_input(
@@ -733,12 +779,25 @@ impl VirtualPool {
             }
         }
         if amount_left != 0 {
-            let next_sqrt_price = get_next_sqrt_price_from_input(
+            let mut next_sqrt_price = get_next_sqrt_price_from_input(
                 current_sqrt_price,
                 config.curve[0].liquidity,
                 amount_left,
                 true,
             )?;
+
+            if next_sqrt_price < config.sqrt_start_price {
+                next_sqrt_price = config.sqrt_start_price;
+                let amount_in = get_delta_amount_base_unsigned(
+                    next_sqrt_price,
+                    current_sqrt_price,
+                    config.curve[0].liquidity,
+                    Rounding::Up,
+                )?;
+                amount_left = amount_left.safe_sub(amount_in)?;
+            } else {
+                amount_left = 0;
+            }
 
             let output_amount = get_delta_amount_quote_unsigned(
                 next_sqrt_price,
@@ -749,10 +808,9 @@ impl VirtualPool {
             total_output_amount = total_output_amount.safe_add(output_amount)?;
             current_sqrt_price = next_sqrt_price;
         }
-        // no need to validate amount_left because if user sell more than what has in quote reserve, then it will be failed when deduct pool.quote_reserve
 
         Ok(SwapAmountFromInput {
-            amount_left: 0,
+            amount_left,
             output_amount: total_output_amount,
             next_sqrt_price: current_sqrt_price,
         })
@@ -895,6 +953,8 @@ impl VirtualPool {
         }
 
         self.update_post_swap(config, old_sqrt_price, current_timestamp)?;
+        // update cached flag
+        self.has_swap = 1;
         Ok(())
     }
 
@@ -933,12 +993,38 @@ impl VirtualPool {
         Ok(())
     }
 
-    pub fn claim_protocol_fee(&mut self) -> (u64, u64) {
-        let token_base_amount = self.protocol_base_fee;
-        let token_quote_amount = self.protocol_quote_fee;
-        self.protocol_base_fee = 0;
-        self.protocol_quote_fee = 0;
-        (token_base_amount, token_quote_amount)
+    pub fn is_first_swap(&self) -> bool {
+        self.has_swap == 0
+    }
+
+    pub fn claim_protocol_base_fee(&mut self, max_amount: u64) -> Result<u64> {
+        // try to claim from trading fees firstly
+        let trading_claimed_fee = self.protocol_base_fee.min(max_amount);
+        self.protocol_base_fee = self.protocol_base_fee.safe_sub(trading_claimed_fee)?;
+        let max_amount = max_amount.safe_sub(trading_claimed_fee)?;
+        // claim from migration fee
+        let migration_claimed_fee = self.protocol_migration_base_fee_amount.min(max_amount);
+        self.protocol_migration_base_fee_amount = self
+            .protocol_migration_base_fee_amount
+            .safe_sub(migration_claimed_fee)?;
+        let total_claimed_fee = trading_claimed_fee.safe_add(migration_claimed_fee)?;
+
+        Ok(total_claimed_fee)
+    }
+
+    pub fn claim_protocol_quote_fee(&mut self, max_amount: u64) -> Result<u64> {
+        // try to claim from trading fees firstly
+        let trading_claimed_fee = self.protocol_quote_fee.min(max_amount);
+        self.protocol_quote_fee = self.protocol_quote_fee.safe_sub(trading_claimed_fee)?;
+        let max_amount = max_amount.safe_sub(trading_claimed_fee)?;
+        // claim from migration fee
+        let migration_claimed_fee = self.protocol_migration_quote_fee_amount.min(max_amount);
+        self.protocol_migration_quote_fee_amount = self
+            .protocol_migration_quote_fee_amount
+            .safe_sub(migration_claimed_fee)?;
+        let total_claimed_fee = trading_claimed_fee.safe_add(migration_claimed_fee)?;
+
+        Ok(total_claimed_fee)
     }
 
     pub fn claim_partner_trading_fee(
@@ -1038,7 +1124,7 @@ impl VirtualPool {
         self.migration_fee_withdraw_status.bitand(mask) == 0
     }
     pub fn update_withdraw_migration_fee(&mut self, mask: u8) {
-        self.migration_fee_withdraw_status = self.migration_fee_withdraw_status.bitxor(mask)
+        self.migration_fee_withdraw_status = self.migration_fee_withdraw_status.bitxor(mask);
     }
 
     pub fn get_migration_progress(&self) -> Result<MigrationProgress> {
@@ -1091,6 +1177,11 @@ impl VirtualPool {
         self.creation_fee_bits = self
             .creation_fee_bits
             .bitxor(PROTOCOL_CREATION_FEE_CLAIMED_MASK)
+    }
+
+    pub fn save_protocol_liquidity_migration_fee(&mut self, base_amount: u64, quote_amount: u64) {
+        self.protocol_migration_base_fee_amount = base_amount;
+        self.protocol_migration_quote_fee_amount = quote_amount;
     }
 }