+++ tests/create_virtual_pool_metadata.tests.ts
@@ -117,6 +117,9 @@ describe("Create virtual pool metadata", () => {
         frequency: 0,
       },
       poolCreationFee: new BN(0),
+      migratedPoolBaseFeeMode: 0,
+      migratedPoolMarketCapFeeSchedulerParams: null,
+      enableFirstSwapWithMinFee: false,
       curve: curves,
     };
     const params: CreateConfigParams<ConfigParameters> = {