+++ tests/token_update_authority.tests.ts
@@ -133,6 +133,9 @@ describe("Create pool with token2022", () => {
         numberOfPeriods: 0,
         frequency: 0,
       },
+      enableFirstSwapWithMinFee: false,
+      migratedPoolBaseFeeMode: 0,
+      migratedPoolMarketCapFeeSchedulerParams: null,
     };
     let params: CreateConfigParams<ConfigParameters> = {
       payer: partner,